Brings code from Keera Breakout for mobile. Refs #10.
This commit brings code from Keera Breakout for iOS/Android, but using a
strange mixture of architectural patterns that borrows from Escape. It
compiles, but currently does not run (segfaults!).

It also shows that graph initialisation, which leads to resource loading,
produces results that need to be used by the renderer (and passed on). Keera
Breakout used a rendering monad to keep and pass that information along.