Stores game ideas & improvements in a new document.
Puts ideas from README and issue #67 in one document called
'help-wanted.md'.

Explanation:
We opened a huge issue (#67) for ZuriHac2018. Afterwards we agreed to
clean this, since this issue is too big and the information could be
stored in a better way.

Note:
The link in README to 'help-wanted.md' hopefully works after a merge to
master.

Closes #67.