Bump version bounds of Yampa and Bearriver. Refs #104.
Haskanoid has an upper limit on the version of Yampa/bearriver of <0.14,
which does not include any of the last two releases. To keep it up to
date with Yampa and bearriver we should increase that to 0.15.

This commit bumps the upper bound on both libraries to 0.15.

Since these packages implement a similar API and are released in sync
with the same version numbers, it makes sense that we bump their version
bounds at the same time.