Update dist in travis config file to Bionic. Refs #105.
Builds on the CI server are failing because the default image uses
xenial, which is no longer supported by Postgresql (and needed for an
apt-get update).

This commit updates the Travis CI config file to select bionic as the
distribution. Because the Haskell image doesn't accept a dist value, we
remove the language field altogether and use an environment matrix to
configure the selected versions of GHC and cabal to use.