Enable CI builds of branches used by our development process. Refs #107.
The travis file is set up to only compile master and develop, but none
of the other branches we are regularly using for development. This
results in an awkward where we have to activate the build on that branch
but then remove the last commit before merging.

This commit enables building any branches that begin with the names
develop, hotfix or release.